Fix "Missing keyboard destroy on unmount" #2522 bug
Added an additional useEffect with an empty dependency array to the 'Keyboard' and 'KeyboardModern' components. This useEffect's return function is called whenever the react component is unmounted and is used here to ensure that destroy() is called on Keyboard object, thus cleaning up any event handlers the Keyboard object may have installed. It also changes the initRef value back to current in case the component is somehow remounted which may happen in certain React versions when running with StrictMode enabled.